chore: refactor dedup package
