+++ services/dedup/types/types.go
@@ -1,0 +1,6 @@
+package types
+
+type KeyValue struct {
+	Key   string
+	Value int64
+}
